requirements.txt
anthropic>=0.25.0
playwright>=1.44.0
chromadb>=0.5.0
python-docx>=1.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
tenacity>=8.0.0
rich>=13.0.0
click>=8.1.0
